Add Sphinx documentation
Currently, the documentation contains the README, all classes or
functions with docstrings, and a new FAQ that contains some
information provided by Eric Hunsberger in #131.

In order to render well and link to external projects' documentation,
some changes were made to several docstrings. With these changes,
the documentation builds without warnings with Sphinx 1.5.2.